ptxdist-import-keys: call cs_append_ca_from_uri for RAUC and HAB SRK roles
The CA consists of the certificate in both cases. This simplifies CA
generation for RAUC which is then installed into the rootfs. In case of
the HAB SRK roles the ptxd_make_imx_habv4_gen_table_impl helper can
simply call cs_get_ca for each role for the srktool call instead of
extract-cert and openssl.

Signed-off-by: Bastian Krause <bst@pengutronix.de>